Gps Settings

+++ lib/controllers/gps.dart
@@ -2,5 +2,5 @@ import 'package:flutter/material.dart';
 
 class GpsController {
   String method = 'distance';
-  int units = 10;
+  double units = 10;
 }
